Simplified time stepping
